refactor logic to useEntryFrom hook
